Remove 8 self-typed casts on file-scope globals
Each cast restates the global's own declared type, so the load already
delivers it:
  - gWeirdMenuSaveTimer is s8 (engine/56 56.c:34) - three sites,
    including the (s8) wrapped around the assignment expression, whose
    value type is already the assigned-to s8.
  - gAndrossProjectileForwardStep / gAndrossAimedProjectileSpeed are int
    (dll_02BC_andross.h:189,192).
  - lbl_803DCC44 is s32 (two sites in objprint_dolphin.c).
  - lbl_803DCF60 is s8 (track_dolphin.c:74).

Only globals were touched. Casts on LOCALS are left alone throughout -
a local is not truncated into its register, so a cast there forces a
real extension; that is why (f32)(s32)noiseFrameCount on the same lines
in objprint_dolphin.c keeps its cast.

All four objects byte-identical (objcopy -R .comment, then
objdump -M gekko -s -t -r); ninja all_source and ninja clean, main.dol OK.

+++ src/dlls/engine/56/56.c
@@ -105,7 +105,7 @@ int WeirdUnusedMenu_run(void)
         {
             saveGame_save();
         }
-        if ((f32)(s8)(gWeirdMenuSaveTimer = ((f32)gWeirdMenuSaveTimer + timeDelta)) >= 120.0f)
+        if ((f32)(gWeirdMenuSaveTimer = ((f32)gWeirdMenuSaveTimer + timeDelta)) >= 120.0f)
         {
             gWeirdMenuPhase = 0;
             gWeirdMenuWidgetWork.widgetFlagsA = (u16)(gWeirdMenuWidgetWork.widgetFlagsA & ~WIDGET_FLAG_SAVING);
